+++ CHANGELOG.md
@@ -4,7 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [4.1.0] - 2018-10-28
+### Added
+- Whirlpool hash algorithm
+- 'Prepend salt' toggle
+- Drupal 7 hash algorithm
+- 'Case-insensitive username' option
+### Fixed
+- Error when 'Display name' not set
+- Encoding of iteration for 'Extended DES (Crypt)'
+- 'Trying to get property of non-object' warning
+
+## [4.0.1] - 2018-08-16
+### Fixed
+- Leftover lines break the admin page
+
+## [4.0.0] - 2018-08-11
 ### Added
 - SHA512 Whirlpool hash algorithm
 - WoltLab Community Framework 2.x hash algorithm
@@ -83,7 +98,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Supported version of ownCloud, Nextcloud: ownCloud 10, Nextcloud 12
 
-[Unreleased]: https://github.com/nextcloud/user_sql/compare/v4.0.0-rc2...develop
+[4.1.0]: https://github.com/nextcloud/user_sql/compare/v4.0.1...v4.1.0
+[4.0.1]: https://github.com/nextcloud/user_sql/compare/v4.0.0...v4.0.1
+[4.0.0]: https://github.com/nextcloud/user_sql/compare/v4.0.0-rc2...v4.0.0
 [4.0.0-rc2]: https://github.com/nextcloud/user_sql/compare/v4.0.0-rc1...v4.0.0-rc2
 [4.0.0-rc1]: https://github.com/nextcloud/user_sql/compare/v3.1.0...v4.0.0-rc1
 [3.1.0]: https://github.com/nextcloud/user_sql/compare/v2.4.0...v3.1.0